Enable ASM helpers and fix 8085 sim execution
- Enable PJVM_ASM_HELPERS for 8085 target builds (-DPJVM_ASM_HELPERS
  in Makefile for core.o and i8085_sim.o)
- Fix static ctx: PJVMCtx on stack triggered LLVM memset intrinsic bug
  (i8 val arg not widened to i16, corrupting size argument and zeroing
  56KB including ROM). Use static PJVMCtx in BSS instead.
- Remove stale PJVM_ASM_HELPERS comment from i8085_sim.c

Results (BubbleSort on i8085 simulator):
  Code size: 24,823B → 23,963B (-860B, -3.5%)
  Runtime:   1,023,783 → 713,528 steps (-30%)
  Cycles:    6,296,559 → 4,731,533 (-25%)

+++ core.c
@@ -273,8 +273,8 @@ void pjvm_parse(uint8_t *data) {
     for (uint8_t i = 0; i < n_methods; i++) {
         m_ml[i] = p[0]; m_ac[i] = p[2]; m_fl[i] = p[3];
         if (version == 0x4B) {
-            m_co[i] = (uint32_t)p[4] | ((uint32_t)p[5] << 8)
-                     | ((uint32_t)p[6] << 16) | ((uint32_t)p[7] << 24);
+            m_co[i] = (uint32_t)((uint32_t)p[4] | ((uint32_t)p[5] << 8)
+                     | ((uint32_t)p[6] << 16) | ((uint32_t)p[7] << 24));
             m_cb[i] = (uint16_t)p[8] | ((uint16_t)p[9] << 8);
             m_vs[i] = p[10]; m_vmid[i] = p[11];
             m_ec[i] = p[12]; m_eo[i] = p[13]; p += 14;
